refactor: split alternativeOrder index into multiple files
