+++ apps/cowswap-frontend/src/common/state/alternativeOrder/atomFactories.ts
@@ -1,0 +1,35 @@
+import { atom, Getter, PrimitiveAtom, SetStateAction, Setter, WritableAtom } from 'jotai'
+
+import { isAlternativeOrderModalVisibleAtom } from './atoms'
+
+export function alternativeOrderAtomGetterFactory<AtomValue>(
+  regular: PrimitiveAtom<AtomValue>,
+  alternative: PrimitiveAtom<AtomValue>
+) {
+  return (get: Getter) => get(get(isAlternativeOrderModalVisibleAtom) ? alternative : regular)
+}
+
+type WritableWithOptionalSetterValue<GetterValue, SetterValue> = WritableAtom<GetterValue, [value: SetterValue], void>
+
+export function alternativeOrderAtomSetterFactory<AtomGetterValue, AtomWriterParamValue>(
+  regular: WritableWithOptionalSetterValue<AtomGetterValue, AtomWriterParamValue>,
+  alternative: WritableWithOptionalSetterValue<AtomGetterValue, AtomWriterParamValue>
+) {
+  return (get: Getter, set: Setter, value: AtomWriterParamValue) => {
+    if (get(isAlternativeOrderModalVisibleAtom)) {
+      set(alternative, value)
+    } else {
+      set(regular, value)
+    }
+  }
+}
+
+export function alternativeOrderReadWriteAtomFactory<AtomType>(
+  regular: WritableWithOptionalSetterValue<AtomType, SetStateAction<AtomType>>,
+  alternative: WritableWithOptionalSetterValue<AtomType, SetStateAction<AtomType>>
+) {
+  return atom(
+    alternativeOrderAtomGetterFactory<AtomType>(regular, alternative),
+    alternativeOrderAtomSetterFactory<AtomType, AtomType>(regular, alternative)
+  )
+}
