+++ apps/cowswap-frontend/src/common/state/alternativeOrder/hooks.ts
@@ -1,0 +1,28 @@
+import { useAtomValue, useSetAtom } from 'jotai'
+import { useCallback } from 'react'
+
+import { Order } from 'legacy/state/orders/actions'
+
+import { ParsedOrder } from 'utils/orderUtils/parseOrder'
+
+import { alternativeOrderAtom, isAlternativeOrderModalVisibleAtom } from './atoms'
+
+export function useIsAlternativeOrderModalVisible() {
+  return useAtomValue(isAlternativeOrderModalVisibleAtom)
+}
+
+export function useHideAlternativeOrderModal() {
+  const setAlternativeOrderAtom = useSetAtom(alternativeOrderAtom)
+
+  return useCallback(() => setAlternativeOrderAtom(null), [setAlternativeOrderAtom])
+}
+
+export function useAlternativeOrder() {
+  return useAtomValue(alternativeOrderAtom)
+}
+
+export function useSetAlternativeOrder() {
+  const setAlternativeOrder = useSetAtom(alternativeOrderAtom)
+
+  return useCallback((order: Order | ParsedOrder) => setAlternativeOrder(order), [setAlternativeOrder])
+}
